add action sheet in delete

+++ UsedGames/UsedGames/ContentView.swift
@@ -10,36 +10,50 @@ import SwiftUI
 struct ContentView: View {
     @ObservedObject var gameStore = GameStore()
     
+    @State var gameToDelete: Game?
+    
     var body: some View {
         List {
             ForEach(gameStore.games) { (game) in
                 GameListItem(game: game)
             }
             .onDelete(perform: { indexSet in
-                gameStore.delete(at: indexSet)
+                self.gameToDelete = gameStore.game(at: indexSet)
             })
             .onMove(perform: { indices,
                 newOffset in
                 gameStore.move(indices: indices, to: newOffset)
             })
         }.padding(EdgeInsets(top: 28, leading: 0, bottom: 0, trailing: 0))
-        .animation(.easeIn, value: gameStore.games)
-        .overlay(
-            VStack {
-                HStack {
-                    EditButton()
+            .animation(.easeIn, value: gameStore.games)
+            .overlay(
+                VStack {
+                    HStack {
+                        EditButton()
+                        Spacer()
+                        Button(action: {
+                            gameStore.createGame()
+                        }, label: {
+                            Text("Add")
+                        }).buttonStyle(BorderlessButtonStyle())
+                    }
+                    .padding()
+                    .background(Color.white.edgesIgnoringSafeArea(.top))
                     Spacer()
-                    Button(action: {
-                        gameStore.createGame()
-                    }, label: {
-                        Text("Add")
-                    }).buttonStyle(BorderlessButtonStyle())
                 }
-                .padding()
-                .background(Color.white.edgesIgnoringSafeArea(.top))
-                Spacer()
+            ).actionSheet(item: $gameToDelete) { (game) -> ActionSheet in
+                ActionSheet(
+                    title: Text("Are you sure?"),
+                    message: Text("You will delete \(game.name)"),
+                    buttons: [
+                        .cancel(),
+                        .destructive(Text("Delete"), action: {
+                            if let indexSet = gameStore.indexSet(for: game) {
+                                gameStore.delete(at: indexSet)
+                            }
+                        })
+                    ])
             }
-        )
     }
 }
 
@@ -66,6 +80,7 @@ struct GameListItem: View {
             Text(NSNumber(value: game.priceInDollas), formatter: numberFormatter)
                 .font(.title2)
                 .foregroundColor(game.priceInDollas > 30 ? .blue : .gray)
+                .animation(nil)
         }
         .padding(.vertical, 4)
     }
